Remove TypedSchemaBuilder
