+++ src/GraphQL/Types/ComplexGraphType.cs
@@ -24,7 +24,7 @@ namespace GraphQL.Types
 
         public IEnumerable<FieldType> Fields
         {
-            get { return _fields; }
+            get => _fields;
             private set
             {
                 _fields.Clear();
